restructuring css
