+++ srcs/requirements/front-end/react-app/js-srcs/Elements/Navbar.jsx
@@ -1,4 +1,5 @@
 import { Link } from "react-router-dom";
+import styles from './Navbar.module.css'
 
 const Navbar = () => {
 	return (
