+++ srcs/requirements/front-end/react-app/js-srcs/Pages/Home.css
@@ -1,7 +1,7 @@
 @import url('https://fonts.googleapis.com/css2?family=KoHo:ital,wght@0,200;0,300;0,400;0,500;0,600;0,700;1,200;1,300;1,400;1,500;1,600;1,700&display=swap');
 /* base styles */
 
-.homepage {
+.home-homepage {
 	margin-left: 10vw;
 	height: 100vh;
 	display: flex;
@@ -9,7 +9,7 @@
 	/* align-items: right; */
 }
 
-.homepage::before {
+.home-homepage::before {
 	content: "";
 	position: absolute;
 	top: 0;
@@ -27,50 +27,50 @@
 	box-shadow: inset 0 0 100px black;
 }
 
-.menu a {
+.home-menu a {
 	text-decoration: none;
 }
 
-.menu {
+.home-menu {
 	display: flex;
 	flex-direction: column;
 	justify-content: center;
 	align-items: left;
 }
 
-.play h1:hover::before {
+.home-play h1:hover::before {
 	content: 'PLAY';
 	position: absolute;
 	filter: blur(7px);
 }
 
-.options h1:hover::before {
+.home-options h1:hover::before {
 	content: 'OPTIONS';
 	position: absolute;
 	filter: blur(7px);
 }
 
-.profile h1:hover::before{
+.home-profile h1:hover::before{
 	content: 'PROFILE';
 	position: absolute;
 	filter: blur(7px);
 }
 
-.menu-item {
+.home-menuItem {
 	display: flex;
 	align-items: center;
 	position: relative;
 }
 
-.menu-item h1:hover{
+.home-menuItem h1:hover{
 	-webkit-text-stroke: 0.03em #0085FF;
 }
 
-.menu-item:hover .rhombus{
+.home-menuItem:hover .rhombus{
 	visibility: visible;
 }
 
-.menu-item h1 {
+.home-menuItem h1 {
 	font-family: "KoHo";
 	color: transparent;
 	font-size: 10vh;
@@ -78,16 +78,16 @@
 	line-height: 0.8;
 }
 
-.rhombus {
+.home-rhombus {
 	width: 8%;
 	height: auto;
 	visibility: hidden;
 }
 
-.rhombus.left {
+.home-rhombus.left {
 	margin-right: 5px;
 }
 
-.rhombus.right {
+.home-rhombus.right {
 	margin-left: 5px;
 }
