+++ srcs/requirements/front-end/react-app/js-srcs/Utils/deletePage.js
