+++ srcs/requirements/front-end/react-app/package.json
@@ -5,7 +5,7 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-	"start": "npx vite --host 0.0.0.0",
+    "start": "npx vite --host 0.0.0.0",
     "dev": "vite",
     "build": "vite build",
     "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "axios": "^1.6.2",
+    "classnames": "^2.3.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.20.1"
